refactor(services): PR-3.3 inference/ + reg/ 子包搬迁 — 6 文件
inference/:
  inference_core.py    → inference/core.py
  inference_daemon.py  → inference/daemon.py
  generate_cache.py    → inference/cache.py
  upscaler.py          → inference/upscaler.py

reg/:
  reg_builder.py    → reg/builder.py
  reg_postprocess.py → reg/postprocess.py
  （analysis.py 抽取留独立 commit）

回归：207 passed（含 inference_core/daemon、generate_cache、upscaler、reg_builder/worker/endpoints）
